Move SerialFolderLocal-type to episodeinfo
