+++ src/main/scala/whodarr/classic/organizer/StoryReorganizer.scala
@@ -1,6 +1,7 @@
 package whodarr.classic.organizer
 
 import os.Path
+import whodarr.classic.episodeinfo.SerialFolderLocal
 import whodarr.classic.episodeinfo.webarchive.{ EpisodeRecognizerImplSimple, SerialFileFilterImpl }
 import whodarr.classic.organizer.webarchive.SerialFilenameConverterImpl
